Compute/requirements.txt
flask
oci
python-dotenv
gunicorn